Added tests to creation of valid properties

+++ app/models/function.rb
@@ -13,21 +13,27 @@ class Function
   validates :name, presence: true
   validates :created_from, presence: true, url: true
 
-  before_save :extract_properties_id, :create_function_properties
+  before_save :create_function_properties
 
 
-  # extract the property id from its uri (for all property)
-  def extract_properties_id
-    properties.each do |property|
-      property['property_id']  = Addressable::URI.parse(property['uri']).basename
-    end
-  end
 
   # create the properties
   def create_function_properties
+    if properties
+      extract_properties_id
       function_properties.destroy_all
-      properties.each do |property|
-        function_properties.build property
-      end
+      properties.each { |property| function_properties.build property }
+    else
+      puts "No properties was sent. Nothing is changing"
+    end 
   end
+
+  private 
+
+    # extract the property id from its uri (for all property)
+    def extract_properties_id
+      properties.each do |property|
+        property['property_id']  = Addressable::URI.parse(property['uri']).basename
+      end
+    end
 end
